% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{auc_pred}
\alias{auc_pred}
\title{auc_pred}
\usage{
auc_pred(beta.hat, V, tt, coeffs, nf = 3)
}
\arguments{
\item{beta.hat}{estimated regression coefficients from the AUC regression model.}

\item{V}{the estimated asymptotic variance from the AUC regression model.}

\item{tt}{a vector of time.}

\item{coeffs}{a vector of covariate values, do not include intercept or observed time, ordered to correspond to the elements in \code{beta.hat} for each covariate}

\item{nf}{a numeric value of 3 or 7 for the number of polynomials. When nf = 7, the vector of polynomials = c(t^{-2},t^{-1},t^{-.5},log(t),t^{0.5},t,t^2). When nf = 3, the vector of polynomials = c(t^{0.5},t, t^2)}
}
\value{
a matrix of 95% confidence intervals for the time-dependent AUC, where the first row represents point estimate, the second row is the lower level, and the third row is the upper level.
}
\description{
This is to estimate the time-dependent AUC with its 95% confidence interval.
}
\examples{
\dontrun{

auc_pred(beta.hat,V.hat, seq(0.05,1,0.01),c((mean_age-min_age)/(max_age - min_age),1,1),3)

} 
}
