% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{covariance_cal}
\alias{covariance_cal}
\title{covariance_cal}
\usage{
covariance_cal(a, b, c, d, e, f, g, h, i, j, k, l, m, n)
}
\arguments{
\item{a}{a vector of regression coefficients from the AUC logistic regression model.}

\item{b}{a vector of biomarker values.}

\item{c}{the \code{rho01} vector from the dataframe returned by the \code{data_crossingdc} function. When there are no continuous variables, use the \code{Iil} vector from the dataframe returned by the \code{data_crossingdc} function.}

\item{d}{the \code{rho02} vector from the dataframe returned by the \code{data_crossingdc} function. When there are no continuous variables, use the \code{1-Iil} vector from the dataframe returned by the \code{data_crossingdc} function.}

\item{e}{A numeric matrix with k rows and n columns. Each row represents a covariate used in the AUC logistic regression model, including the intercept as the first column. The number of columns is equal to the number of rows in the dataframe returned by the \code{data_crossingdc} function.}

\item{f}{the \code{idi - 1} vector from the dataframe returned by the \code{data_crossingdc} function.}

\item{g}{the \code{idl - 1} vector from the dataframe returned by the \code{data_crossingdc} function.}

\item{h}{a numeric matrix with q rows and n columns. Each row represents a covariate used in the missing data logistic regression model for cases, including the intercept as the first column. The number of columns is equal to the number of rows in the dataframe returned by the `data_crossingdc` function}

\item{i}{a numeric matrix with q rows and n columns. Each row represents a covariate used in the missing data logistic regression model for controls in the risk set, including the intercept as the first column. The number of columns is equal to the number of rows in the dataframe returned by the `data_crossingdc` function}

\item{j}{a vector of estimated probability of observing biomarkers among case patients.}

\item{k}{a vector of estimated probability of observing biomarkers among control patients.}

\item{l}{a vector of regression coefficiens from the missing data model.}

\item{m}{a numeric matrix, returned from the \code{pi.gamma.omega.est} function, accessed through \code{pi.gamma.omega.est$composit}.}

\item{n}{a numeric matrix,, returned from the \code{pi.gamma.omega.est} function, accessed through \code{pi.gamma.omega.est$h.hat}.}
}
\value{
A list containing:
\itemize{
  \item \code{sigma1}: The matrix \eqn{\Sigma_1}, used in calculating the asymptotic variance.
  \item \code{sigma2}: The matrix \eqn{\Sigma_2}, also used in calculating the asymptotic variance.
  \item \code{V}: The asymptotic variance for the regression coefficents.
}
}
\description{
This function computes three matrices, \eqn{\Sigma_1},\eqn{\Sigma_2}, and \eqn{V}.  \eqn{V} is the asymptotic variance, defined as \eqn{\Sigma_1^{-1} \Sigma_2 \Sigma_1^{-1}}.
}
