% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{data_crossingdc}
\alias{data_crossingdc}
\title{data_crossingdc}
\usage{
data_crossingdc(V, vt, vc, vm, vxs, pi, cont_vars, discrete_vars, c0)
}
\arguments{
\item{V}{a numeric vector of missing data indicators, V = 1 if observed and V = 0 if missing}

\item{vt}{a numeric vector of observed time}

\item{vc}{a numeric vector of censoring indicator}

\item{vm}{a numeric vector of biomarker values}

\item{vxs}{a covariate matrix used for the AUC logistic regression, do not include intercept or observed time}

\item{pi}{estimated probability of observing biomarker values}

\item{cont_vars}{a vector of column names from vxs representing continuous covariates}

\item{discrete_vars}{a vector of column names from vxs representing discrete covariates}

\item{c0}{A constant value used for the bandwidth h, h = c0*n^(-1/3)}
}
\value{
A dataframe where each row corresponds to a case-control pair at each observed event time.
The dataframe has the following components:
\itemize{
  \item \code{i}: numeric index for cases.
  \item \code{idi}: patient id for cases.
  \item \code{idl}: patient id for controls.
  \item \code{yi}: observed event time for cases.
  \item \code{yl}: observed event time for controls in the risk set of i.
  \item \code{Iil}: indicator for comparison of baseline biomarker values (1 if case biomarker > control biomarker, 0 otherwise).
  \item \code{pii}: estimated probability of observing the biomarker for cases.
  \item \code{pil}: estimated probability of observing the biomarker for controls in the risk set of i.
  \item \code{event_indicator_i}: censoring indicator for cases, this is always 1.
  \item \code{event_indicator_l}: censoring indicator for controls in the risk set of i. 1 if event occurred, 0 if censored.
  \item \code{mi}: biomarker values for cases.
  \item \code{ml}: biomarker values for controls in the risk set of i.
  \item \code{Covariates with "i" suffix}: covariate columns, with names based on the original covariate name and an "i" suffix for case values.
  \item \code{rho01} and \code{rho02}: numeric vectors representing concordant and discordant event values, respectively, when continuous covariates are present.
  \item \code{rhoweight}: Kernel weight vector, calculated with continuous covariates.
}
}
\description{
This is to reshape data for each cause of failure, the result reshaped data is ready for parameter estimation.
}
\examples{
\dontrun{
Y0<-x[,"time_year"]   
C0<-x[,"status"]
M0<-x[,"TAU_bl"]
VXS<-x[,c("age","sex", "race")];
V0 <- dat$V_tau
pi.hat <- est.pi.func(x, c("age","bmi"), c("sex","race") ,"discrete",u, V, futime, status,8, c(0,0,0,0,0,0,0), "L-BFGS-B")$pi.hat
data_crossingdc(V, Y0,C0,M0,VXS, pi.hat, , c("age"),c("sex", "race"), 1.5)

} 
}
