% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo}
\alias{demo}
\title{Demo data}
\format{
## `demo`
A dataframe of 1000 rows and 7 columns:
\describe{
  \item{ID}{Patient ID}
  \item{Y}{Observed event time}
  \item{delta}{Censoring indicator. 0 = censored, 1 = event 1}
  \item{z}{A numeric vector indicating a continuous non-instrument covariate}
  \item{u}{A numeric vector indicating a discrete instrument covariate}
  \item{M}{Baseline biomarker values}
  \item{V}{A missing indicator for biomarker values, V = 1 if observed, V = 0 if missing}
}
}
\usage{
demo
}
\description{
A data used for demo.
}
\keyword{datasets}
