% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{est.pi.func}
\alias{est.pi.func}
\title{est.pi.func}
\usage{
est.pi.func(
  x,
  z_cont,
  z_dis,
  u_type,
  u,
  V,
  futime,
  status,
  q,
  initial,
  method = c("Nelder-Mead", "BFGS", "CG", "L-BFGS-B", "SANN", "Brent")
)
}
\arguments{
\item{x}{a numeric matrix containing the design matrix in the missing data logistic regression model.}

\item{z_cont}{a vector of column names from x representing continuous covariates that are non-instrumental variables. If there is no continuous covariate, set it as NULL.}

\item{z_dis}{a vector of column names from x representing categorical covariates that are non-instrumental variables. If there is no categorical covariate, set it as NULL.}

\item{u_type}{a character string indicating the type of instrumental variable, set to "continuous" or "dicrete".}

\item{u}{a numeric vector representing the instrumental variable.}

\item{V}{a numeric vector of missing data indicators, V = 1 if observed and V = 0 if missing.}

\item{futime}{a numeric vector of observed event time.}

\item{status}{a numeric vector of censoring indicator, status = 1 if event occurs and status = 0 if censored.}

\item{q}{a numeric value indicating the number of estimating equations.}

\item{initial}{a numeric vector of initial values for the parameters in the missing data logistic regression model.}

\item{method}{a character string indicating the optimization method, set to  one of the "Nelder-Mead", "BFGS", "CG", "L-BFGS-B", "SANN","Brent".}
}
\value{
a list, containing the following items:
\itemize{
  \item \code{phi.hat}: estimated regression coefficients from the missing data model.
  \item \code{convergence}: convergence code from the second step optimization.
  \item \code{convergence.message}: convergence message from the second step optimization.
  \item \code{objective}: value of the objective function from the second step optimization.
  \item \code{vcov.hat}: variance-covariance matrix for the parameter estimates in the missing data logistic regression model. 
  \item \code{pi.hat}: estimated probability of observing the biomarker
  \item \code{h.hat}: a numeric matrix with n rows and q columns, each row is \eqn{h(\hat{\phi},y_i)} 
  \item \code{composit}:this is \eqn{{(\Gamma^T{\Omega}^{-1}\Gamma)}^{-1}\Gamma^T{\Omega}^{-1}}.
}
}
\description{
This is to perform two-step GMM estimation for the missing data logistic regression model using optim().
}
\examples{
\dontrun{
est.pi.func(x, c("age","bmi"), c("sex","race") ,"discrete",u, V, futime, status,8, c(0,0,0,0,0,0,0), "L-BFGS-B")

} 
}
