\name{tAUCMNAR-package}
\alias{tAUCMNAR-package}
\alias{tAUCMNAR}
\docType{package}
\title{
\packageTitle{tAUCMNAR}
}
\description{
\packageDescription{tAUCMNAR}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{tAUCMNAR}

Maintainer: \packageMaintainer{tAUCMNAR}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = tAUCMNAR)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{tAUCMNAR}}
%\section{Documentation Index}{\packageIndices{tAUCMNAR}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
